Adjusted to review, moved the 0 effort check to GripperCommand
