+++ Gems/ROS2/Code/Source/Utilities/JointUtilities.h
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
+ * SPDX-License-Identifier: Apache-2.0 OR MIT
+ *
+ */
+
+#pragma once
+#include <AzCore/Component/Entity.h>
+
+namespace ROS2::Utils
+{
+    AZStd::string GetJointName(AZ::EntityId entityId);
+}
